first version of questionnaire page ready

+++ src/main/java/tools/CarbonCalculator.java
@@ -69,7 +69,7 @@ final class CarbonCalculator {
      * @param litres , the amount of litres of heating oil consumed.
      * @return carbon emissions in kg
      */
-    public static double heatingOilEmissions(double litres) {
+    public static double heatingOilEmissions(int litres) {
         return litres * 2.96;
     }
